requirements.txt
uagents>=0.23.6
uagents-core>=0.4.0
openai>=1.0.0
sendgrid>=6.9.0
requests>=2.31.0
python-dotenv>=1.0.0
cosmpy>=0.9.0
